requirements.txt
azarashi
dearpygui
gpiozero
smbus2
pyserial
